% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyLMM-package.R
\docType{package}
\name{dyLMM-package}
\alias{dyLMM}
\alias{dyLMM-package}
\title{dyLMM: Dyadic Linear Mixed Models}
\description{
Fit dyadic linear mixed models (LMM/MLM/HLM) including ICC models and
actor-partner interdependence models (APIM) for indistinguishable and
distinguishable dyads.
}
\section{Data wrangling}{

\itemize{
\item \code{\link[=build_composites]{build_composites()}} - Aggregate indicators to sum or mean composites
\item \code{\link[=wide_to_long]{wide_to_long()}} - Reshape wide dyad data to long format
}
}

\section{Model fitting}{

\itemize{
\item \code{\link[=fitICC]{fitICC()}} - Intraclass correlation (one-way random effects)
\item \code{\link[=fitAPIMindist]{fitAPIMindist()}} - Indistinguishable actor-partner interdependence model
\item \code{\link[=fitAPIMdist]{fitAPIMdist()}} - Distinguishable actor-partner interdependence model
}
}

\section{Helpers}{

\itemize{
\item \code{\link[=getICC]{getICC()}} - Extract ICC from fitted ICC model
\item \code{\link[=outputLMMTab]{outputLMMTab()}} - Reproducible parameter table (tibble or gt)
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jsakaluk/dyLMM/}
  \item Report bugs at \url{https://github.com/jsakaluk/dyLMM/issues}
}

}
\keyword{internal}
