% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitICC.R
\name{fitICC}
\alias{fitICC}
\title{Fit an Intraclass Correlation (ICC) Model}
\usage{
fitICC(data, y, dyad_id = "dyad_id", ...)
}
\arguments{
\item{data}{Data frame in long format (two rows per dyad).}

\item{y}{Character. Name of the outcome column.}

\item{dyad_id}{Character. Name of the dyad identifier column. Default \code{"dyad_id"}.}

\item{...}{Further arguments passed to \code{\link[nlme:gls]{nlme::gls()}}.}
}
\value{
A fitted \code{gls} object from \code{nlme::gls}.
}
\description{
Fits a model with compound symmetry to estimate the intraclass correlation
(ICC) for dyadic data. Uses \code{\link[nlme:gls]{nlme::gls()}}. Use \code{\link[=getICC]{getICC()}} or \code{\link[=getRho]{getRho()}} to
extract the ICC (rho) from the fitted model.
}
\examples{
\dontrun{
library(dySEM)
dvn <- scrapeVarCross(dat = commitmentQ, x_order = "spi", x_stem = "sat.g",
  x_delim1 = ".", x_delim2 = "_", distinguish_1 = "1", distinguish_2 = "2",
  verbose = FALSE)
dat_comp <- build_composites(dat = commitmentQ, dvn = dvn)
dat_long <- wide_to_long(dat_comp)
fit <- fitICC(data = dat_long, y = "y")
getICC(fit)
}
}
